chore: reorganize files and folders
