+++ .babelrc
@@ -1,11 +1,7 @@
 {
     "presets": ["@babel/preset-env"],
     "plugins": [
-        ["module-resolver",  { "root": ["./"],
-            "alias": {
-                "threeExamples": "./src/Renderer/ThreeExtended/jsPatched/"
-            }
-        } ],
+        ["module-resolver",  { "root": ["./"] } ],
         ["babel-plugin-inline-import", {
             "extensions": [
                 ".json",
