+++ src/Converter/textureConverter.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import Feature2Texture from '../Renderer/ThreeExtended/Feature2Texture';
+import Feature2Texture from './Feature2Texture';
 
 const textureLayer = (texture) => {
     texture.generateMipmaps = false;
