+++ src/Core/Prefab/Planar/PlanarTileBuilder.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import OBB from '../../../Renderer/ThreeExtended/OBB';
+import OBB from '../../../Renderer/OBB';
 import Coordinates from '../../Geographic/Coordinates';
 import Extent from '../../Geographic/Extent';
