+++ test/unit/layeredmaterialnodeprocessing.js
@@ -3,7 +3,7 @@ import assert from 'assert';
 import { updateLayeredMaterialNodeImagery } from '../../src/Process/LayeredMaterialNodeProcessing';
 import TileMesh from '../../src/Core/TileMesh';
 import Extent from '../../src/Core/Geographic/Extent';
-import OBB from '../../src/Renderer/ThreeExtended/OBB';
+import OBB from '../../src/Renderer/OBB';
 import Layer from '../../src/Layer/Layer';
 import Source from '../../src/Source/Source';
 import { STRATEGY_MIN_NETWORK_TRAFFIC } from '../../src/Layer/LayerUpdateStrategy';
