+++ test/unit/obb.js
@@ -5,7 +5,7 @@ import Extent from '../../src/Core/Geographic/Extent';
 import BuilderEllipsoidTile from '../../src/Core/Prefab/Globe/BuilderEllipsoidTile';
 import PlanarTileBuilder from '../../src/Core/Prefab/Planar/PlanarTileBuilder';
 import TileGeometry from '../../src/Core/TileGeometry';
-import OBB from '../../src/Renderer/ThreeExtended/OBB';
+import OBB from '../../src/Renderer/OBB';
 
 const max = new THREE.Vector3(10, 10, 10);
 const min = new THREE.Vector3(-10, -10, -10);
